requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-bootstrap5==2.1.2
idna==3.3
oauthlib==3.2.0
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9

python-dateutil~=2.8.2
matplotlib~=3.5.3
django-model-utils~=4.2.0